% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RnaSeqTutorials-package.R
\docType{package}
\name{RnaSeqTutorial05 package}
\alias{RnaSeqTutorial05}
\alias{RnaSeqTutorial05-package}
\alias{RnaSeqTutorial05 package}
\title{RnaSeqTutorials package}
\description{
A simple description of the RnaSeqTutorial05 package
}
\section{Tutorials}{

This is the fifth in a series of tutorials
\itemize{
\item\code{05_ensemble_gene_set_enrichment_analysis} a tutorial on ensemble gene set enrichment analysis
}
}

\examples{
	\dontrun{
	learnr::run_tutorial("05_ensemble_gene_set_enrichment_analysis", package = "RnaSeqTutorial05")
	}
}
\seealso{
The vignette
}
\author{
Nicolas Delhomme [aut,cre]
}
\keyword{internal}
\keyword{package}
